:sparkles: Componentizacao da linha da tabela
