+++ src/pages/InstitutionsPage/index.jsx
@@ -1,4 +1,5 @@
 import { Container } from './styles'
+import InstitutionRow from '../../components/InstitutionRow'
 import image from '../../assets/foto.png'
 
 const InstitutionsPage = () => {
@@ -14,24 +15,11 @@ const InstitutionsPage = () => {
           </tr>
         </thead>
         <tbody>
-          <tr>
-            <td>
-              <img src={image} alt="" />
-            </td>
-            <td>Instituição Criança Top</td>
-            <td>
-              <p> R$ 3.500,00/ R$ 4.0000,00 </p>
-              <div className='progress-bar'>
-                <div className='progress'></div>
-                <div className='not-progress'></div>
-              </div>
-            </td>
-            <td>
-              <button>
-                Detalhes
-              </button>
-            </td>
-          </tr>
+          <InstitutionRow image={image}/>
+          <InstitutionRow image={image}/>
+          <InstitutionRow image={image}/>
+          <InstitutionRow image={image}/>
+          <InstitutionRow image={image}/>
         </tbody>
       </table>
     </Container>
